+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
  - Changed transition form type. It requires the transition handler instead of the transition now
  - Change interface of TransitionFormBuilder
+ - Only pass one transition form builder to the transition form type
 
 ## [2.0.0-rc3] - 2019-09-02